refactor(audience): keep delete on dimension audience only
A filtered audience is just a stored filter tree pointing at a
dimension audience's graduates — deleting it has no observable effect.
Move ``delete`` off ``RapidataAudienceBase`` onto ``RapidataAudience``
so the slim ``RapidataFilteredAudience`` doesn't advertise a no-op.
